fix(core): fix frame buffer size on disc2
[no changelog]

+++ core/embed/trezorhal/stm32u5/xdisplay/stm32u5a9j-dk/display_fb.c
@@ -31,10 +31,10 @@
 
 // Physical frame buffers in internal SRAM memory
 __attribute__((section(".fb1")))
-ALIGN_32BYTES(uint32_t physical_frame_buffer_0[PHYSICAL_FRAME_BUFFER_SIZE]);
+ALIGN_32BYTES(uint8_t physical_frame_buffer_0[PHYSICAL_FRAME_BUFFER_SIZE]);
 
 __attribute__((section(".fb2")))
-ALIGN_32BYTES(uint32_t physical_frame_buffer_1[PHYSICAL_FRAME_BUFFER_SIZE]);
+ALIGN_32BYTES(uint8_t physical_frame_buffer_1[PHYSICAL_FRAME_BUFFER_SIZE]);
 
 // The current frame buffer selector at fixed memory address
 // It's shared between bootloaders and the firmware
